doc/requirements.txt
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinx-copybutton==0.5.2
sphinxcontrib-plantuml==0.27
sphinxcontrib-jquery==4.1
Pygments==2.17.2
docutils==0.20.1
PyYAML==6.0.1
Pillow==10.3.0
graphviz==0.20.1
packaging==21.3
docopt==0.6.2
